Refactor record_move (#13)
Dynamically allocate move_record array to conserve space.

The initial size of move_record is set to 5 because a minimum
of 5 moves is required to determine the winner.

And then in a 4 x 4 board, if both players aim to win, the
game can always conclude within 5 moves. Therefore,
it is currently unclear what the optimal size for resizing.