Exand label and beautify table
Orignially labels at left side and bottom could be non-alphanumeric or
even unprintable when board size is larger than 10. Also, original input is
capped in [0, 9] for rows and [a, z] for colomns.

This commit expands lebal in two aspects: label display and input.
Column label display now supports from 'A' to 'ZZ' (676 columns), and
row label display supports from '1' to '999'. I suggest not to support futher
since it would exceeds the limitation of human eyes. So printing is hard
coded under this limit.

Any cell in the table now can be selected by typing in the corresponding
label value. Input is decoupled with label display so it is still
functional when board size exceeds the limit of label. Note that the
variables to store parsed x, y input is int so baord size is capped to
INT_MAT - 1, or the maximum size of table. To support futher, I suggest
make the table hierarchical like virtual memory to avoid unsupported
board size. Nicer error messages are also provided now.